refactor(conversation): render the transcript on assistant-ui
`#/conversation` now runs on `@assistant-ui/react` through an external-store
runtime: the library owns the composer, the scroll viewport and the run state,
while `AppShell` keeps owning the transcripts.

That ownership split is the reason for the external-store runtime specifically.
A reply can arrive over SSE while this view is unmounted, and a turn can outlive
the request that started it (#983), so the message array cannot move into the
library — the external store is the one runtime that takes it as an input.

The transport is untouched: still `POST …/chat` on the Rust host and the company
event stream, with the same three send outcomes (resolved, detached, failed)
reported to the shell. Nothing here speaks a data-stream protocol.

Carried over unchanged: board cards (#246), card dismissal and its 404 path
(#984), per-reply step timelines, the live SSE tool timeline, the working
indicator, and the in-flight steer strip (#111). The runtime is handed no
`onCancel` — abandoning the last reply is not the action an operator wants
here; steering a named run is, and only that strip can name them.

Grouping and day separators move into a pure `decorate` pass, because
assistant-ui renders one message at a time and a line's neighbours are no longer
visible at render. `test/unit/conversation-model.test.ts` covers them; before
this they were observable only on screen.

Both lockfiles are updated: `npm ci` in the Desktop job reads
package-lock.json, and the Console job validates pnpm-lock.yaml against
package.json on pnpm 11 and on the pnpm 10 floor (issue #858).
